requirements.txt
streamlit
polars
pyarrow
pyyaml
